requirements.txt
streamlit
pytesseract
pymupdf  
pillow 
python-dotenv